requirements.txt
flask==3.0.3
gunicorn==22.0.0
numpy==1.26.4
pandas==2.2.2
google-api-python-client==2.149.0
google-auth==2.35.0
google-cloud-storage==2.18.2
